src/llm/requirements.txt
# LLM Query Interface Requirements
# Install with: pip install -r requirements.txt --break-system-packages

# Core LLM SDKs
openai>=1.0.0                    # GPT-4 and DeepSeek R1
anthropic>=0.40.0               # Claude Sonnet 4
google-generativeai>=0.8.0      # Gemini 2.0 Flash
together>=1.0.0                 # Llama 3.3 70B

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Progress tracking
tqdm>=4.65.0

# Dataset loading
ucimlrepo>=0.0.3                # UCI ML Repository datasets
scikit-learn>=1.3.0             # For data preprocessing
